Introduce Apollo KV Cache interface and generic Keyv wrapper (#93)
We've gone back and forth on making the switch to keyv as the cache
implementation in Apollo Server. We've decided we want to own the API via an
interface but still easily support keyv usage via a wrapper.

This PR introduces both:

* the interface (AS3's existing `KeyValueCache` interface, so AS4's cache type
  is no longer a breaking change) + accompanying implementations
  `InMemoryLRUCache`, `PrefixingKeyValueCache`, and `ErrorsAreMissesCache`
* the `ApolloKeyv` wrapper for `Keyv`s. This wrapper supports the basic methods
on `Keyv` (`get`, `set`, `delete`). The `keyv` wrapper can be used like so:

```
new ApolloServer({
  cache: new ApolloKeyv(new Keyv("redis://...")),
  //...
});
```

Keyv supports a number of adapters which we have no business maintaining (like
Redis!), so this allows to remain backwards-compatible with existing caches and
cache consumers (like plugins!) while also being able to support great caches
without reinventing the wheel.

Related: apollographql/apollo-server#6045