Add throwIfAborted & reason to AbortSignal
I added `// @ts-expect-error` to the abortable fetcher tests because `undici` defines the options param to `addEventListener` with `once?: boolean` instead of `once?: boolean | undefined;`.

That makes the test work, but not sure if clients will also need to add that, which is not ideal. I tried setting `exactOptionalPropertyTypes` but that would definitely require clients using undici to set it in their tsconfig.

https://www.typescriptlang.org/tsconfig#exactOptionalPropertyTypes