New package: @apollo/cache-control-types
This package will move to the apollo-server repo once v4.0.0 is
released, but we're putting it here for now so we can do a non-alpha
release.

This will allow us to (among other things) remove the use of
`apollo-server-types` in `@apollo/subgraph`, and remove the `declare
module` in AS4 that would conflict with the AS3 version if both versions
end up installed in a single TS project.

Note that the `CacheAnnotation` type is not moved into this package;
instead, it will become an internal implementation detail of the cache
control plugin.