keyvaluecache: add PrefixesAreUnnecessaryForIsolationCache
This allows software like Apollo Server plugins to "prefix" their
provided cache by default, but still allow you to disable prefixes by
passing in a specially "tagged" cache.

Part of apollographql/apollo-server#6742. Will close that issue once AS
is updated to throw if one of these caches is provided to it.
Also part of apollographql/datasource-rest#19.