Enforce `DataLoader` contract (#260)
`KeyvAdapter` currently provides an invalid function to `DataLoader`. Though the types
don't indicate it, `keyv.get([])`(the plural/array overload) is capable of returning
`undefined` in the event that the underlying `Store.getMany()` returns `undefined`.
When this happens, `KeyvAdapter` violates `DataLoader`s expectations (and it throws
an error).

We can defend against this by handling the `undefined` case and returning an array
of `keys.length` filled with `undefined`.