Change generatePersistedQueryIdsFromManifest to async manifest; fix Promises
This function seems as likely as the other to benefit from async
manifest loads.

The test I added here failed at first because of the unhandled
rejection. (I'm not positive why the
createPersistedQueryManifestVerificationLink error test didn't fail for
the same reason.) Putting an empty catch on these Promises seems legit,
as we will not completely lose track of the rejection.