Log warning when no operations are found and add number of operations matched during generation (#348)
When generating a persisted query manifest file, its possible that no operations were found during manifest generation, which could be due to a misconfiguration of the glob pattern used to search for operations. Currently the user has to run the command, then open the manifest file to see if operations were added which can be pretty cumbersome. This updates the terminal output to list the number of matched operations. If none were found, a warning is displayed as well.